build: deal with newer glibc vs. FORTIFY_SOURCE
Recent glibc (for whatever stupid reason) decided to issue a #warning
when FORTIFY_SOURCE is requested but not possible, due to compilation
without optimization.  But when debugging sources, we don't want our
use of -Werror to kill compilation just because glibc wants to be
chatty; when debugging, we don't care about source fortification.
* configure.ac (FORTIFY_SOURCE): Make conditional on optimization.

Signed-off-by: Eric Blake <eblake@redhat.com>