build: get missing from automake, not gnulib
Automake recently changed 'missing' in a backwards-incompatible
manner (that is, the 'missing' from newer automake won't meet the
usage pattern of a makefile generated by an older automake).
While this has since been fixed, the consensus was that gnulib
should not be tracking 'missing' in the first place, since it is
too heavily tied to automake internals and not independently
useful.  We can prepare for this by getting 'missing' as part of
our bootstrap process instead of linking into gnulib.

See https://lists.gnu.org/archive/html/automake-patches/2012-06/msg00154.html
for more discussion.

* build-aux/missing: Break link to gnulib.
* .gitignore: Reflect this.

Signed-off-by: Eric Blake <eblake@redhat.com>