build: silence gcc warning
Our next gnulib update will automatically turn on -Wformat=2, which
overrides our attempt to avoid -Wformat-nonliteral.  But for newer
gcc, we don't need to blindly ignore the warning; instead, we can
target just the audited code while leaving the warning to apply to
the rest of the code.  And for older gcc, we just don't use the
warning (that is, warnings are a moving target, and it is fine if
we only target newer gcc for full warnings).

* configure.ac (M4_cv_gcc_pragma_push_works): New test.
(-Wformat-nonliteral): Disable for older gcc.
* src/format.c (expand_format): Mark our only uses of nonliteral
format as being safe.

Signed-off-by: Eric Blake <eblake@redhat.com>