maint: update to latest gnulib
Upstream gnulib has added several improvements:

- the regex module allows for rational range interprtation (that
is, guaranteeing that the regex "[a-z]" evaluates to the same 26
bytes as in ASCII, even when in a non-C locale).  Technically,
m4 isn't yet using non-C locales, so I don't know if this will
make much difference, but it can't hurt to include
- various build fixes that I found while smoke-testing on
alternative platforms
- bump the copyright year to 2013 (I'll do m4 next)

* gnulib: Update to latest, for rational range regex, build fixes,
copyright update.
* m4/gnulib-cache.m4: Regenerate.
* NEWS: Document this.

Signed-off-by: Eric Blake <eblake@redhat.com>