build: fix large diversions on mingw
Gnulib's clean-temp module defaults to using _O_TEMPORARY on mingw,
which has the effect of deleting a file the first time it is closed.
This resulted in a testsuite failure in 146.diversions, with m4
complaining that it could not create a temporary file for holding a
diversion.  In reality, m4 was trying to rename the file, but
because mingw does not allow renaming an open file, we have to close
it first - but closing it first deleted the file.

For now, we will just work around the gnulib issue by disabling the
use of _O_TEMPORARY.

This was probably introduced in commit da5f6db (1.4.13), although
my testing environment was painfully slow to confirm this, so I
did not call out a 'regression since xyz' in the NEWS entry.

* gl/lib/clean-temp.c.diff: New file.

Signed-off-by: Eric Blake <eblake@redhat.com>