build: improve default build for non-gcc
Dagobert Michelsen reported that in a setup that uses git to
track patches on top of the distributed tarball, we ended up
having configure defaults that did not play nicely with the
Sun Studio 12.3 compiler.  Gnulib has fixed the problems with
detecting (un-)supported compiler flags, and additionally we
can use a better witness file of whether we are likely to be
a developer where it is worth enabling warnings by default.

* gnulib: Update to latest gnulib.
* configure.ac (gl_WARN_ADD): Drop flags now guaranteed by gnulib.
(--enable-gcc-warnings): Set default based on a tarball-only file,
rather than whether .git exists.
* THANKS: Update.

Signed-off-by: Eric Blake <eblake@redhat.com>