doc: fix line-wrapped macro definitions
The html rendering of the manual ate a comma and inserted an
awkward line wrap for the "example" macro prototype:
https://www.gnu.org/software/m4/manual/m4-1.4.17/html_node/Manual.html

As documented in autoconf, use of @c does not work for macros
which are designed for inline use within @deffn and friends.
http://git.savannah.gnu.org/gitweb/?p=autoconf.git;a=commitdiff;h=a357718

* doc/m4.texi: Drop bogus @c.
Reported by Marijn Schouten, fix by Patrice Dumas copied from
autoconf code base.

Signed-off-by: Eric Blake <eblake@redhat.com>