symtab: better handling of macro stacks
I ran into a scenario where running a program took 22s with the
default -H509, but less than a second with -H517 [1].  The culprit?  A
collision between 'stack' and 'substr' in the default hash table size
caused lookups for substr to get progressively slower as pushdef stack
got deeper.  This is easy enough to fix, and may also make it easier
to dynamically grow the hashtable.

[1] https://lists.gnu.org/archive/html/bug-m4/2021-04/msg00000.html

* src/m4.h (struct symbol): Add stack member.
* src/symtab.c (lookup_symbol): Separate stack from bucket list.
(symtab_print_list): Update traversal to match.
* src/freeze.c (produce_frozen_state): Likewise.
(reverse_symbol_list): Reverse stack, not bucket.