maint: port to macOS 11.2.3 (arm64)
* m4/gnulib-cache.m4: Add fopen-gnu, replacing cloexec and fopen.
Avoid getopt-posix-tests, since they are not needed for m4
and currently fail on macOS 11.2.3 (arm64).
* src/builtin.c (m4_incr, m4_decr): Avoid undefined behavior
on integer overflow that causes tests to fail on macOS.
* src/debug.c (debug_set_output):
* src/output.c (m4_tmpfile, m4_tmpopen):
* src/path.c (m4_fopen):
Use GNU fopen with "e" rather than set_cloexec_flag.  This is
simpler, and works around a Gnulib bug on macOS with fopen
being replaced by rpl_fopen sometimes but not other times.
* src/freeze.c (produce_frozen_state): Use GNU fopen with "e";
no need to expose the fd to subprocesses.