syscmd: Allow commands with leading - or +
As POSIX recently pointed out[1], anything with semantics like
system() or popen() should be passing "--" between "-c" and the user's
string, in case the user intends to execute a utility beginning with
'-' or '+'.  POSIX recommends that users should not name files
beginning with '-', but does not have a similar discouragement against
files beginning with '+'.  In particular, if your /bin/sh is bash and
you want m4 to fork to a script named "+O" rather than incorrectly
printing a list of shopt settings, this patch is essential.  If you
need to be portable to older m4, you can always prepend a space in
your arguments to syscmd().

[1] https://www.austingroupbugs.net/view.php?id=1440

* src/builtin.c (m4_syscmd, m4_esyscmd): Pass "--" to sh prior to
user's string.