maint: Silence compiler false positive
gcc 14.2.1 warned that output_text() could be calling memcpy() with
output_cursor NULL and length non-zero, after a call to
make_room_for().  But this is a false positive: if output_cursor is
NULL after make_room_for(), it is because the diversion switched over
to an active output_file.

* src/output.c (output_text): Add assertion to silence gcc.