eval: Drop obsolete use of = in eval
Ever since 1.4.8b in 2007, we have been giving deprecation warnings
to anyone using the non-POSIX spelling of `=' for an eval equality
comparison (so that an eventual m4 2.0 can use it for assignment
instead).  It's been long enough to pull the plug on this crutch.

* src/eval.c (enum eval_token): Reorder to be in precedence order,
with values assigned in groups of 10. Drop ASSIGN.
(equality_term): Drop support for ASSIGN.
* NEWS: Document this.
* doc/m4.texi (Eval): Likewise.