changeword: Fix coredump on invalid regex
The docs are clear that the only valid use of the experimental
changeword is to pass a regex where all prefixes of a desired
acceptable word are also accepted.  If this constraint is not met, and
the regex can match a longer string but not the one-byte prefix of
that string, then the fastmap of that regex will still be set on that
byte (as it could be a valid anchor to try searching for a longer
match), but when re_search() then fails to match, we are left with
garbage in regs.start.

The docs even had an example where this constraint is not met, but
because of the way the test was written, the first macro parsed after
changeword was "dnl", which populated regs with something that happens
to work on the next attempt to parse with failure to match "f".  But
the test added here demonstrates that without a prior regex match, if
the first byte after changeword is in the fastmap but fails to parse,
then regs.start will still be NULL and crash m4.  And even when m4
didn't crash, it's still better to only rely on regs after re-running
the regex on the largest string that did match, rather than whatever
is left in regs after the first failure to match a one-byte-longer
string.

I _really_ want to get rid of changeword.  It slows things down, and
is a nightmare to maintain.  And the fact that NO ONE reported this
regression introduced in 2008 (because most distros wisely refuse to
build with --enable-changeword) means it won't be missed.  But
removing a feature, even experimental, should be done for 1.6, not
1.4.x.

* NEWS: Document the bug fix.  Also a whitespace fix.
* cfg.mk (old_NEWS_hash): Run 'make update-NEWS-hash'.
* m4.texi (Changeword): Test for the bug.
* input.c (word_start) [ENABLE_CHANGEWORD]: Revert commit cde8ed62,
but this time use a static array rather than malloc'd pointer.
(set_word_regexp): Populate word_start.
(peek_token): Use it.
(next_token): Likewise, and don't use regs.start[1] if regex did not
have \(\) grouping.