build: update to latest gnulib
Rather than trying to cherry-pick all gnulib changes on branch-1.4,
this just updates to the latest (see commit 2f445aef) then fixes ALL
of the fallout, as learned by referencing a number of other commits.
Done this way since it is otherwise impossible to build this branch,
in an effort to revive it.  With this in place, it should be easier to
do more targetted forward ports of other commits.

* gnulib: Update to latest.
* m4/gnulib-cache.m4: Regenerate.  Change pipe to spawn-pipe.
* .gitignore: Cater to changes in build-aux.
* build-aux/missing: Drop; newer automake doesn't need it.
* doc/m4.texinfo: Rename to doc/m4.texi, and re-encode in UTF-8.
Cater to quotearg changing `' to '' in messages throughout
(@setcontentsaftertitlepage): Drop.
* examples/null.err: Cater to quotearg changes.
* doc/Makefile.am (info_TEXINFOS): Track rename.
* checks/Makefile.in (stamp-checks): Likewise.
* configure.ac (AC_PREREQ, AM_INIT_AUTOMAKE): Bump to newer baseline.
(AC_TYPE_SIGNAL): Drop deprecated macro.
* src/builtin.c (m4_syscmd, m4_esyscmd): Track execute and
create_pipe_in changes.
(mkstemp_helper): Track obstack changes.
* src/debug.c (trace_flush): Likewise.
* src/input.c (next_token): Likewise.
* src/m4.h (includes): Drop obsolete "verror.h".
* src/output.c (m4_tmpfile, m4_tmpopen): Track fopen_temp changes.
* src/symtab.c (lookup_symbol): Track hash_delete changes.