changeword: Remove this experimental feature
Changeword is seldom-tested and bug-prone, and slows m4 down (distros
don't enable it).  No one has asked on the mailing lists for it to
work better, or submitted patches.  M4 2.0 might be able to offer
changesyntax as an alternative that is not as slow, and users are
still free to stick with 1.4.x if they don't want to upgrade to 1.6.
But since we've been warning about it being experimental for years
now, it's time to rip it out rather than trying to keep patching it.

* NEWS: Mention the removal.
* README: Drop mention of changeword.
* configure.ac: No longer support --enable-changeword.
* doc/m4.texi (Top, Operation modes, Quoting Arguments, Changequote)
(Changecom, Using frozen files): Drop mention of the feature.
(Changeword): Delete this section.
(Frozen file format): Mention how changeword from 1.4.x is handled.
* examples/null.m4: Drop use of changeword.
* examples/null.out: Regenerate.
* examples/null.err: Likewise.
* src/m4.h [ENABLE_CHANGEWORD]: Delete all code related to changeword.
* src/builtin.c: Likewise.
* src/input.c: Likewise.
* src/m4.c: Likewise.