doc: Be more specific about regexp syntax
As evidenced by recent gnulib traffic [1], GNU Emacs regexp syntax has
diverged over time, to the point that RE_SYNTAX_EMACS==0 is no longer
accurate: modern Emacs has since enabled a\{2\} interval repetition,
as well as [[:alpha:]] char classes, neither of which is supported in
current m4.  However, for back-compat reasons, we cannot blindly
change m4 1.4.x away from syntax 0 even if it is no longer Emacs
syntax.  Worse, at least Autoconf 2.72 has instances of regex where
both "{" and "\{" are intended to match a literal "{".  Enabling
intervals could cause regex that compile now to fail to compile and
cause a warning, which is a change that can only be done on a major
version bump to 1.6.  So for now, just document the limitations.  [1]
https://lists.gnu.org/archive/html/bug-gnulib/2025-04/msg00064.html

* doc/m4.texi (regexp): Document highlights for users that don't want
to chase the link, and call out intentional lack of newer features
in contrast to what Emacs now supports.
(patsubst): Refer to regexp, rather than Emacs.