maint: Ensure stable timestamp for manual
Git likes to set mtime of files to the point where the working
directory is checked out, rather than the point where the last content
of the file was committed.  But Automake likes to populate the
generated doc/version.texi based on the mtime of doc/m4.texi.  Any
setup that uses a git checkout of m4 to build a new tarball will thus
get different contents in the manual if checked out on a different
date, breaking reproducible builds unless we take measures to
guarantee that the mtime matches the time of the last commit.

* configure.ac (st_touch): New code.
* THANKS: Update.
Suggested by Simon Josefsson, after a report by Santiago Vila:
https://lists.gnu.org/archive/html/bug-m4/2025-04/msg00052.html