debugmode: Add support for ? query of current flags
I'm hoping to add a debugmode(+r) to toggle whether to warn on
suspicious regex constructs.  But I couldn't do it in m4 1.4.x, in
part because debugmode() warns on unrecognized flags.  It will be
easier to add new flags in the future if it is also possible to query
the set of flags known by m4, as well as the current state of those
flags.  And the easiest way to do that was to have debugmode(?) output
a string containing both the set and cleared flags, which means the
parser has to accept that as well.  For back-compat with 1.4.x,
debugmode(-) is once again synonymous with clearing the default flags
but leaving others in place.  And the earlier change to make 'm4 -d'
behave like 'm4 -d+adeq' but 'debugmode()' behave like
'debugmode(aeqt)' is too complex to maintain; it's simpler to have the
empty argument set the default rather than add the default.

* doc/m4.texi (Debugging options): Tweak example to match code.
(Debugmode): Mention qindir, missed in previous patch. Document new
ability to support + and - in same string, and new ? for query.
* src/builtin.c (m4_debugmode): Handle new "?" case.
* src/m4.h (debug_dump): New declaration.
* src/debug.c (debug_dump): New function.
(debug_set): New helper, factored out of...
(debug_decode): ...here. Restore 1.4.x compatibility with parsing
"", "-", and "+".  Support more than one "-" and "+" in a string.
* NEWS: Document this.