doc: Improve curry example
Classic lambda calculus defines the Curry function to only apply one
additional argument; but it is just as easy in m4 to curry an
arbitrary number of fixed arguments coupled with an arbitrary number
of extra arguments.  It is also worth documenting how to provide a
name to a curried function.

* examples/curry.m4: Accept multiple extra arguments, and improve
documentation.
* doc/m4-texi (Composition): Reflect it into the manual.
Reported-by: Nikolaos Chatzikonstantinou
https://lists.gnu.org/archive/html/m4-discuss/2025-05/msg00056.html