eval: Reject incomplete base prefix
POSIX is clear that strtol("0x") is an error (there must be digits
after the explicit base); and other m4 implementations that rely on
strtol for eval have rejected "eval(0x)".  Although it can be argued
that rejecting something that used to silently work as 0 in m4 1.4.x
might break existing scripts, we had never documented it as a feature
and the risk is low.

Meanwhile, encountering "eval(08)" is always going to be an error, but
the code was previously stopping the parse at the first out-of-range
digit, tokenizing as "0" followed by "8", which produces a semantic
error when 8 is found when an operator was expected.  However, the
error message is nicer if we instead trigger the same new error as for
an incomplete base ["invalid number"] rather than the old way ["bad
expression in eval (excess input)"]; a future patch may then be able
to call attention to which substring of the overall expression was
actually the bad operator or bad number, instead of printing the
entire expression.  Compare to how bash parses all alphanumerics after
a leading digit as the token it complains about, even if it detects
the error earlier in the token.

* src/eval.c (MIN_PREC): New define to avoid a magic number to
parse_expr.
(BADNUM, INVALID_NUMBER): New enum values.  Shuffle values on
other eval_tokens to keep math consistent.
(eval_lex): Return BADNUM on failed numeric parse.
(primary, parse_expr, evaluate): Handle invalid numbers.
* doc/m4.texi (Eval): Test it.
* NEWS: Document the bug fix.

(cherry picked from commit 730e37ab488659fd1fe117054381babee4183eee)